This is an ugly hack to improve the performance in the computation of the subcell limiting formula.

We avoid a lot of multiplications and additions with 0. As a result, the computation of the DG staggered fluxes goes from `2.30μs` to `1.82μs` . However, the non-conservative fluxes become less readable...

This is the *new* performance summary of the test referenced in https://github.com/trixi-framework/Trixi.jl/pull/1670:
```
────────────────────────────────────────────────────────────────────────────────────────
                Trixi.jl                        Time                    Allocations
                                       ───────────────────────   ────────────────────────
           Tot / % measured:                 115s /  97.2%           3.01GiB /  99.9%

 Section                       ncalls     time    %tot     avg     alloc    %tot      avg
 ────────────────────────────────────────────────────────────────────────────────────────
 rhs!                           4.87k    92.9s   83.2%  19.1ms   11.5KiB    0.0%    2.43B
   volume integral              4.87k    73.6s   66.0%  15.1ms   2.20KiB    0.0%    0.46B
     calcflux_fhat!             19.9M    36.3s   32.5%  1.82μs     0.00B    0.0%    0.00B
     calcflux_fv!               19.9M    21.6s   19.4%  1.08μs     0.00B    0.0%    0.00B
     calcflux_antidiffusive!    19.9M    8.70s    7.8%   437ns     0.00B    0.0%    0.00B
     ~volume integral~          4.87k    7.05s    6.3%  1.45ms   2.20KiB    0.0%    0.46B
   interface flux               4.87k    11.9s   10.7%  2.45ms     0.00B    0.0%    0.00B
   prolong2interfaces           4.87k    3.33s    3.0%   685μs     0.00B    0.0%    0.00B
   surface integral             4.87k    2.01s    1.8%   413μs     0.00B    0.0%    0.00B
   reset ∂u/∂t                  4.87k    1.31s    1.2%   270μs     0.00B    0.0%    0.00B
   Jacobian                     4.87k    609ms    0.5%   125μs     0.00B    0.0%    0.00B
   ~rhs!~                       4.87k   22.4ms    0.0%  4.61μs   9.33KiB    0.0%    1.96B
   prolong2mortars              4.87k   3.01ms    0.0%   618ns     0.00B    0.0%    0.00B
   mortar flux                  4.87k   1.65ms    0.0%   339ns     0.00B    0.0%    0.00B
   prolong2boundaries           4.87k   1.37ms    0.0%   282ns     0.00B    0.0%    0.00B
   boundary flux                4.87k    353μs    0.0%  72.6ns     0.00B    0.0%    0.00B
   source terms                 4.87k    120μs    0.0%  24.7ns     0.00B    0.0%    0.00B
 a posteriori limiter           4.87k    12.3s   11.1%  2.54ms   1.15MiB    0.0%     248B
   blending factors             4.87k    7.25s    6.5%  1.49ms   1.15MiB    0.0%     248B
     positivity                 4.87k    6.65s    6.0%  1.37ms   1.15MiB    0.0%     248B
     ~blending factors~         4.87k    592ms    0.5%   122μs      752B    0.0%    0.15B
   ~a posteriori limiter~       4.87k    5.10s    4.6%  1.05ms      752B    0.0%    0.15B
 I/O                              165    3.36s    3.0%  20.3ms   2.17GiB   72.0%  13.4MiB
   save solution                  164    3.35s    3.0%  20.4ms   2.17GiB   71.9%  13.5MiB
   get element variables          164   7.49ms    0.0%  45.7μs   1.42MiB    0.0%  8.89KiB
   ~I/O~                          165   1.82ms    0.0%  11.0μs    379KiB    0.0%  2.30KiB
   save mesh                      164   62.2μs    0.0%   379ns     0.00B    0.0%    0.00B
 calculate dt                   1.62k    2.11s    1.9%  1.30ms     0.00B    0.0%    0.00B
 analyze solution                  18    896ms    0.8%  49.8ms    862MiB   28.0%  47.9MiB
 ────────────────────────────────────────────────────────────────────────────────────────
```